More damage caps

+++ Server/Models/Monsters/NetherworldGate.cs
@@ -106,7 +106,7 @@ namespace Server.Models.Monsters
             return;
         }
 
-        public override int Attacked(MapObject attacker, int power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             return 0;
         }
